Shop: Rename module to `store`
