+++ rare/components/tabs/store/constants.py
