+++ rare/components/tabs/store/game_widgets.py
@@ -6,7 +6,7 @@ from PyQt5.QtGui import QFont
 from PyQt5.QtNetwork import QNetworkAccessManager
 from PyQt5.QtWidgets import QWidget, QVBoxLayout, QLabel, QHBoxLayout
 
-from rare.components.tabs.shop.shop_models import ImageUrlModel
+from rare.components.tabs.store.shop_models import ImageUrlModel
 from rare.ui.components.tabs.store.wishlist_widget import Ui_WishlistWidget
 from rare.utils.extra_widgets import ImageLabel
 from rare.utils.misc import icon
