+++ rare/components/tabs/store/search_results.py
