+++ rare/components/tabs/store/shop_widget.py
